+++ registry/lib/pages/checkbox_page.dart
@@ -160,16 +160,16 @@ class _WithLabel extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const DefaultCheckbox(
-      vm: CheckboxModel(
-        name: "withLabel",
-        items: [
-          DpItem(
+    return DefaultCheckbox(
+      decorationBuilder: (themeVm) => CheckboxDecoration(
+        themeVm,
+        value: CheckboxValue(themeVm, name: "withLabel", items: [
+          const DpItem(
               id: "1",
               title: "Accept terms and conditions",
               subtitle:
                   "You agree to our Terms of Service and Privacy Policy."),
-        ],
+        ]),
       ),
     );
   }
@@ -180,15 +180,12 @@ class _Disabled extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const DefaultCheckbox(
-      vm: CheckboxModel(
-        name: "disabled",
-        enabled: false,
-        items: [
-          DpItem(id: "1", title: "Accept terms and conditions"),
-        ],
-      ),
-    );
+    return DefaultCheckbox(
+        decorationBuilder: (themeVm) => CheckboxDecoration(themeVm,
+            value: CheckboxValue(themeVm, name: "disabled", items: [
+              const DpItem(id: "1", title: "Accept terms and conditions"),
+            ]),
+            state: CheckboxState(themeVm, isDisabled: true)));
   }
 }
 
@@ -206,27 +203,32 @@ class _Form extends StatelessWidget {
         DefaultForm(
           vm: formModel,
           child: DefaultCheckbox(
-            vm: CheckboxModel(
-              title: "Sidebar",
-              subtitle: "Select the items you want to display",
-              disabled: const [
-                "home",
-                "apps",
-              ],
-              autovalidateMode: AutovalidateMode.onUserInteraction,
-              helperText: "You have to select at least one item",
-              name: "termsForm",
-              orientation: OptionsOrientation.vertical,
-              validator: FormBuilderValidators.required(
-                  errorText: 'You have to select at least one item'),
-              onChanged: print,
-              items: const [
-                DpItem(id: "recents", title: "Recents"),
-                DpItem(id: "home", title: "Home"),
-                DpItem(id: "apps", title: "Applications"),
-                DpItem(id: "settings", title: "Settings"),
-                DpItem(id: "about", title: "About"),
-              ],
+            decorationBuilder: (themeVm) => CheckboxDecoration(
+              themeVm,
+              value: CheckboxValue(themeVm,
+                  name: "termsForm",
+                  disabledItems: const [
+                    "home",
+                    "apps",
+                  ],
+                  autovalidateMode: AutovalidateMode.onUserInteraction,
+                  validator: FormBuilderValidators.required(
+                      errorText: 'You have to select at least one item'),
+                  items: const [
+                    DpItem(id: "recents", title: "Recents"),
+                    DpItem(id: "home", title: "Home"),
+                    DpItem(id: "apps", title: "Applications"),
+                    DpItem(id: "settings", title: "Settings"),
+                    DpItem(id: "about", title: "About"),
+                  ]),
+              action: CheckboxAction(themeVm, onChanged: print),
+              child: CheckboxChild(
+                themeVm,
+                title: "Sidebar",
+                subtitle: "Select the items you want to display",
+                helperText: "You have to select at least one item",
+                orientation: OptionsOrientation.vertical,
+              ),
             ),
           ),
         ),
@@ -254,21 +256,22 @@ class _Card extends StatelessWidget {
         return CardDecoration(context,
             size: CardSize(context, padding: const EdgeInsets.all(16)),
             child: CardChild(context,
-                custom: const CardCustom(
+                custom: CardCustom(
                     widget: DefaultCheckbox(
-                  vm: CheckboxModel(
-                    name: "settingsField",
-                    orientation: OptionsOrientation.vertical,
-                    items: [
-                      DpItem(
-                        id: "settings",
-                        title: "Use different settings for my mobile devices",
-                        subtitle:
-                            "You can manage your mobile notifications in the mobile settings page.",
-                      ),
-                    ],
-                  ),
-                ))));
+                        decorationBuilder: (themeVm) => CheckboxDecoration(
+                            themeVm,
+                            value: CheckboxValue(themeVm,
+                                name: "settingsField",
+                                items: [
+                                  const DpItem(
+                                      id: "settings",
+                                      title:
+                                          "Use different settings for my mobile devices",
+                                      subtitle:
+                                          "You can manage your mobile notifications in the mobile settings page.")
+                                ]),
+                            child: CheckboxChild(themeVm,
+                                orientation: OptionsOrientation.vertical))))));
       },
     );
   }